grpc-health-check: remove grpc dependency
Requires requiring the module with `grpc` as an argument,
which requires a major version bump.

Fixes #926
